Extended launch-configuration.json to all commands and listeners
This is to allow for later features to turn on and off these parts using the config to save progress. Such features could be settings in the settings menu that control certain bot elements. These could edit the config with the user having the ability to later easily delete it and reset all they have changed or to just edit it as they like.